+++ .eslintrc.js
@@ -276,13 +276,7 @@ module.exports = {
 					...restrictedSyntax,
 					...restrictedSyntaxComponents,
 					// Temporary rules until we're ready to officially deprecate the bottom margins.
-					...[
-						'BaseControl',
-						'ComboboxControl',
-						'RangeControl',
-						'SearchControl',
-						'SelectControl',
-					].map( ( componentName ) => ( {
+					...[ 'BaseControl' ].map( ( componentName ) => ( {
 						selector: `JSXOpeningElement[name.name="${ componentName }"]:not(:has(JSXAttribute[name.name="__nextHasNoMarginBottom"]))`,
 						message:
 							componentName +